+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 
+### Changed
+
+- Use `segmentToken` instead of `getSegment`.
+
 ## [0.11.1] - 2020-03-27
 
 ### Fixed